+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 - fix: deal with the threads being initialized in the `process_frame` callback with a `threading.Lock`
+- refactor: use a thread-pool instead of creating a new thread for each render
 
 ## Version 0.12.3